Move all gem files under a directory with the same plugin name #13
