+++ lib/hammer_cli_foregit/file_manager.rb
