+++ lib/hammer_cli_foregit/foreman/api.rb
