+++ lib/hammer_cli_foregit/git_manager.rb
