+++ lib/hammer_cli_foregit/talk_commands.rb
